+++ js/main.js
@@ -1,2 +1,3 @@
-import {ads} from './data.js';
+import {createAds} from './data.js';
+const ads = createAds();
 console.log(ads);
